% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{collapse}
\alias{collapse}
\title{Collapse a commondata table by aggregating the most end denominator.}
\usage{
collapse(table, group_cols = c("geoName", "provenanceDomain"), FUN = sum)
}
\arguments{
\item{table}{required, the commondata table.}

\item{group_cols}{optional, the column names for applying the "group by"
operation. The columns "geoNames" and "provenanceDomain" by default.}

\item{FUN}{optional, the function to aggregate the numerical data. Sum by
default.}
}
\value{
A new commondata table
}
\description{
Collapse a commondata table by aggregating the most end denominator.
}
