% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{combine}
\alias{combine}
\title{Combine two commondata data tables by aligning the table columns properly.
The data are then grouped based on the group_cols and then aggregated using
the function FUN parameters. The result is a new combined commondata data
table.}
\usage{
combine(d1, d2, group_cols = c("geoName", "provenanceDomain"), FUN = sum)
}
\arguments{
\item{d1}{required, the first commondata data table.}

\item{d2}{required, the second commondata data table.}

\item{group_cols}{optional, the column names for the group operation.
"geoNames" and "provenanceDomain" columns are used by default.}

\item{FUN}{optional, the function to aggregate the numerical data. The sum
operation is used by default.}
}
\value{
A new commondata data table
}
\description{
Combine two commondata data tables by aligning the table columns properly.
The data are then grouped based on the group_cols and then aggregated using
the function FUN parameters. The result is a new combined commondata data
table.
}
\examples{
# Combine the count data between male and female populations in Palo Alto
combine(count_female_population(PALO_ALTO_CITY_CA), count_male_population(PALO_ALTO_CITY_CA))
}
