% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_household_income_by_age.R
\name{count_household_income_by_age}
\alias{count_household_income_by_age}
\title{Return the householder count per income bracket and per age group for the
given the geographical names (i.e., zip codes, state codes, state names,
county names) and theobservation year period.}
\usage{
count_household_income_by_age(
  geo_names,
  location_type = c(NA, "zip", "county", "state"),
  start_year = 2011,
  end_year = 2018,
  year = NA
)
}
\arguments{
\item{geo_names}{required, vector of string(s) of geographical names.}

\item{location_type}{optional, string indicating the location type of the
geographical names. NA by default.}

\item{start_year}{optional, integer indicating the start year of observation.
2011 by default.}

\item{end_year}{optional, integer indicating the end year of observation.
2018 by default.}

\item{year}{optional, integer indicating a single year of observation.
This parameter overrides the start_year and end_year parameters when
it is not NA, such that start_year=year and end_year=year. NA by default.}
}
\value{
A named list with each list item is a data frame containing the
household income observations of each region per income bracket. The data
frame is identified by the observation year and age group.
}
\description{
Return the householder count per income bracket and per age group for the
given the geographical names (i.e., zip codes, state codes, state names,
county names) and theobservation year period.
}
\examples{
zips <- c("94035","94039","94040","94041","94042","94043")

# Count the householders per income and age group in the specified ZIP codes
count_household_income_by_age(zips)

# Count the householders from 2012 to 2015
count_household_income_by_age(zips, start_year=2012, end_year=2015)

# Count the householders in 2012
count_household_income_by_age(zips, year=2012)

# Count the householders in the state of California
count_household_income_by_age(c("California")) # State name
count_household_income_by_age(c("CA")) # State code
count_household_income_by_age(c("06")) # 2-digit state FIPS code
}
