% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_population.R
\name{count_male_population}
\alias{count_male_population}
\title{Return the male human population count per age group for the given the
geographical names (i.e., zip codes, state codes, state names, county names)
and theobservation year period.}
\usage{
count_male_population(geo_names, start_year = 2011, end_year = 2018, year = NA)
}
\arguments{
\item{geo_names}{required, vector of string(s) of geographical names}

\item{start_year}{optional, integer indicating the start year of observation.
2011 by default.}

\item{end_year}{optional, integer indicating the end year of observation.
2018 by default.}

\item{year}{optional, integer indicating a single year of observation.
This parameter overrides the start_year and end_year parameters when
it is not NA, such that start_year=year and end_year=year. NA by default.}
}
\value{
A named list with each list item is a data frame containing the
male population count of each region per age group. The data frame is
identified by the observation year.
}
\description{
Return the male human population count per age group for the given the
geographical names (i.e., zip codes, state codes, state names, county names)
and theobservation year period.
}
\examples{
zips <- c("94035","94039","94040","94041","94042","94043")

# Count the male population in the specified ZIP codes
count_male_population(zips)

# Count the male population from 2012 to 2015
count_male_population(zips, start_year=2012, end_year=2015)

# Count the male population in 2012
count_male_population(zips, year=2012)

# Count the male population in the state of California
count_male_population(c("California")) # State name
count_male_population(c("CA")) # State code
count_male_population(c("06")) # 2-digit state FIPS code
}
