% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_count_covid19_incident.R
\name{cumulative_count_covid19_incident}
\alias{cumulative_count_covid19_incident}
\title{Return the cumulative count of COVID-19 incident per case category for the
given the geographical names (i.e., zip codes, state codes, state names,
county names) and the observation date period.}
\usage{
cumulative_count_covid19_incident(
  geo_names,
  location_type = c(NA, "zip", "county", "state"),
  start_date = "2020-01-01",
  end_date = "2020-03-31",
  date = NA
)
}
\arguments{
\item{geo_names}{required, vector of string(s) of geographical names.}

\item{location_type}{optional, string indicating the location type of the
geographical names. NA by default.}

\item{start_date}{optional, date indicating the start date of observation.
"2020-01-01" by default.}

\item{end_date}{optional, date indicating the end date of observation.
"2020-03-31" by default.}

\item{date}{optional, date indicating a single date of observation.
This parameter overrides the start_date and end_date parameters when
it is not NA, such that start_date=date and end_date=date. NA by default.}
}
\value{
A named list with each list item is a data frame containing the
cumulative count observations of COVID-19 incident per case category.
The data frame is identified by the observation date.
}
\description{
Return the cumulative count of COVID-19 incident per case category for the
given the geographical names (i.e., zip codes, state codes, state names,
county names) and the observation date period.
}
\examples{
states <- c("Alabama","California","New York","Pennsylvania","Washington")

# Count the cumulative count of COVID-19 incidents in the specified states
cumulative_count_covid19_incident(states)

# Count the cumulative count of COVID-19 incidents from mid-June until today
cumulative_count_covid19_incident(states, start_date="2020-06-30", end_date=Sys.Date())

# Count the cumulative count of COVID-19 incidents 10 days ago from today
cumulative_count_covid19_incident(states, date=Sys.Date()-10)
}
