% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_person_age_by_sex.R
\name{median_person_age}
\alias{median_person_age}
\title{Return the median age of the population for the given the geographical
names (i.e., zip codes, state codes, state names, county names, or school
district) and the observation year period.}
\usage{
median_person_age(
  geo_names,
  location_type = c(NA, "zip", "county", "state", "school"),
  start_year = 2011,
  end_year = 2018,
  year = NA
)
}
\arguments{
\item{geo_names}{required, vector of string(s) of geographical names.}

\item{location_type}{optional, string indicating the location type of the
geographical names. NA by default.}

\item{start_year}{optional, integer indicating the start year of observation.
2011 by default.}

\item{end_year}{optional, integer indicating the end year of observation.
2018 by default.}

\item{year}{optional, integer indicating a single year of observation.
This parameter overrides the start_year and end_year parameters when
it is not NA, such that start_year=year and end_year=year. NA by default.}
}
\value{
A named list with each list item is a data frame containing the
median age of the population of each region. The data frame is identified
by the observation year and the sex category.
}
\description{
Return the median age of the population for the given the geographical
names (i.e., zip codes, state codes, state names, county names, or school
district) and the observation year period.
}
\examples{
zips <- c("94035","94039","94040","94041","94042","94043")

# Get the median age in the specified ZIP codes
median_person_age(zips)

# Get the median age from 2012 to 2015
median_person_age(zips, start_year=2012, end_year=2015)

# Get the median age in 2012
median_person_age(zips, year=2012)

# Get the median age in California state
median_person_age(c("California")) # State name
median_person_age(c("CA")) # State code
median_person_age(c("06")) # 2-digit state FIPS code
}
