% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolkit.R
\name{summarize}
\alias{summarize}
\title{Summarize multiple data frames by aggregating the numerical data of same
grouping columns.}
\usage{
summarize(..., group_cols = c("geoName", "provenanceDomain"), FUN = sum)
}
\arguments{
\item{...}{required, the list of data frames.}

\item{group_cols}{optional, the column names for applying the "group by"
operation. The columns "geoNames" and "provenanceDomain" by default.}

\item{FUN}{optional, the function to aggregate the numerical data. Sum by
default.}
}
\value{
A new data frame
}
\description{
Summarize multiple data frames by aggregating the numerical data of same
grouping columns.
}
